Moved inline funcs to header
