+++ hw-1/schedule/schedule_private.h
@@ -2,6 +2,8 @@
 #define HW1_SCHEDULE_PRIVATE_H
 
 #include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
 #include <time.h>
 
 int AddBeginTime(FILE* file, time_t* startTime);
@@ -17,7 +19,14 @@ int GetYear(FILE* file, int* year);
 int GetGroup(FILE* file, int* group);
 int AddLesson(FILE* file, Lessons* lesson);
 void PrintLesson(Lesson const* lesson);
-inline void copy_string(char src[], size_t len, char** dst);
-inline size_t calculate_group_year_index(int const group, int const year);
+
+inline void copy_string(char src[], size_t len, char** dst) {
+  src[len - 1] = (src[len - 1] == '\n') ? '\0' : src[len - 1];
+  *dst = (char*)malloc(len * sizeof(char));
+  memcpy(*dst, src, len * sizeof(char));
+}
+inline size_t calculate_group_year_index(int const group, int const year) {
+  return (year - 1) + (year - 1) * (group - 1);
+}
 
 #endif  // HW1_SCHEDULE_PRIVATE_H
